stm32f4 - add configuration option to set flash latency
According to the STM32F401 and STM32F405 reference manual the
flash needs wait states when the CPU is clocked fast (see the
section named "Relation between CPU clock frequency and
Flash memory read time").

My STM32F405 based board won't boot with the CPU running at 84MHz
unless the flash latency is set to at least 1 which is how I
discovered this.  If I want to follow the manufacturer's
specifications I should set the flash latency to 2.

The reset value is 0 which means that most of the STM32F4xx boards
in the tree that run the CPU at 84MHz are overclocking the internal
flash and are lucky to work at all.

To fix this, add a configuration option to set the flash latency.
The option has a default value of 0 so to not affect existing boards.

Note that 0 is probably wrong for many boards and ought to be set
to something higher in the the board configuration.

Change-Id: Iba2534776fd1e6a90aa68f837045983f9a1b5753
Signed-off-by: Christer Weinigel <christer@weinigel.se>